(Qt) Combine non-generic options into qt/options/nongeneric.cpp - all these Qt files being combined now led to a nearly 1MB size reduction in retroarch's binary
