+++ Makefile.common
@@ -494,19 +494,8 @@ ifeq ($(HAVE_QT), 1)
           ui/drivers/qt/qt_playlist.o \
           ui/drivers/qt/qt_downloads.o
     ifeq ($(HAVE_MENU), 1)
-       OBJ += ui/drivers/qt/options/achievements.o \
-              ui/drivers/qt/options/audio.o \
-              ui/drivers/qt/options/generic.o \
-              ui/drivers/qt/options/input.o \
-              ui/drivers/qt/options/latency.o \
-              ui/drivers/qt/options/network.o \
-              ui/drivers/qt/options/osd.o \
-              ui/drivers/qt/options/playlists.o \
-              ui/drivers/qt/options/recording.o \
-              ui/drivers/qt/options/saving.o \
-              ui/drivers/qt/options/ui.o \
-              ui/drivers/qt/options/user.o \
-              ui/drivers/qt/options/video.o
+       OBJ += ui/drivers/qt/options/nongeneric.o \
+              ui/drivers/qt/options/generic.o
     endif
 
    MOC_HEADERS += ui/drivers/ui_qt.h \
