+++ griffin/griffin_cpp.cpp
@@ -59,20 +59,9 @@ UI
 #include "../ui/drivers/qt/qt_downloads.cpp"
 #ifdef HAVE_MENU
 #include "../ui/drivers/qt/options/generic.cpp"
-#include "../ui/drivers/qt/options/video.cpp"
-#include "../ui/drivers/qt/options/audio.cpp"
-#include "../ui/drivers/qt/options/saving.cpp"
-#include "../ui/drivers/qt/options/osd.cpp"
-#include "../ui/drivers/qt/options/input.cpp"
-#include "../ui/drivers/qt/options/latency.cpp"
-#include "../ui/drivers/qt/options/playlists.cpp"
-#include "../ui/drivers/qt/options/user.cpp"
-#include "../ui/drivers/qt/options/recording.cpp"
-#include "../ui/drivers/qt/options/ui.cpp"
-#include "../ui/drivers/qt/options/achievements.cpp"
-#include "../ui/drivers/qt/options/network.cpp"
-#include "../ui/drivers/qt/moc_settingswidgets.cpp"
+#include "../ui/drivers/qt/options/nongeneric.cpp"
 #include "../ui/drivers/qt/options/moc_options.cpp"
+#include "../ui/drivers/qt/moc_settingswidgets.cpp"
 #endif
 #include "../ui/drivers/moc_ui_qt.cpp"
 #include "../ui/drivers/qt/moc_coreinfodialog.cpp"
